HittableList to separate module
