+++ src/hittable/mod.rs
@@ -11,8 +11,9 @@ use rand::distributions::Uniform;
 use rand::distributions::uniform::UniformFloat;
 use rand::RngCore;
 
-pub use sphere::*;
 pub use aarect::*;
+pub use list::*;
+pub use sphere::*;
 
 use crate::aabb::AABB;
 use crate::material::{Lambertian, Material};
@@ -23,6 +24,7 @@ use crate::vec::{Axis, V3};
 
 mod sphere;
 mod aarect;
+mod list;
 
 #[derive(Copy, Clone)]
 pub struct Hit<'a> {
@@ -45,42 +47,6 @@ impl<'a> Hit<'a> {
 pub trait Hittable: Debug + Sync {
     fn hit(&self, ray: &Ray, dist_min: f64, dist_max: f64) -> Option<Hit>;
     fn bounding_box(&self, t_min: f32, t_max: f32) -> Option<AABB> { None }
-}
-
-#[derive(Debug)]
-pub struct HittableList {
-    objects: Vec<Box<dyn Hittable>>,
-    aabb: AABB,
-}
-
-impl HittableList {
-    pub fn new(objects: Vec<Box<dyn Hittable>>) -> HittableList {
-        let aabb = (|| {
-            let mut aabbs = objects.iter().flat_map(|o| o.bounding_box(0.0, 1.0));
-            let first = aabbs.next()?;
-            Some(aabbs.fold(first, |a, b| a + b))
-        })();
-        HittableList { objects, aabb: aabb.unwrap() }
-    }
-}
-
-impl Hittable for HittableList {
-    fn hit(&self, ray: &Ray, dist_min: f64, dist_max: f64) -> Option<Hit> {
-        self.objects
-            .iter()
-            // todo[performance]: try enabling again after implementing heavier object
-//            .filter(|h| h.bounding_box(ray.time, ray.time)
-//                .map(|aabb| aabb.hit(ray, dist_min, dist_max))
-//                .unwrap_or(true)
-//            )
-            .map(|h| h.hit(ray, dist_min, dist_max))
-            .filter_map(std::convert::identity)
-            .min_by(|s, o| s.dist.partial_cmp(&o.dist).unwrap())
-    }
-
-    fn bounding_box(&self, t_min: f32, t_max: f32) -> Option<AABB> {
-        Some(self.aabb)
-    }
 }
 
 pub trait Instance {
